refactor: extract RDP canvas helpers
